fix: move components
